+++ Project1/Project1/main.cpp
@@ -1,8 +1,8 @@
 // main.cpp : Ce fichier contient la fonction 'main'. L'exécution du programme commence et se termine à cet endroit.
-// Exécuter le programme : Ctrl+F5 ou menu Déboguer > Exécuter sans débogage
-// Déboguer le programme : F5 ou menu Déboguer > Démarrer le débogage
+// Exécuter le programme : Ctrl+F5 ou menu Déboguer > Exécuter sans débogage
+// Déboguer le programme : F5 ou menu Déboguer > Démarrer le débogage
 
-#include "CMatrice.h"
+#include "CFichier.h"
 
 #include <iostream>
 
@@ -11,8 +11,8 @@ int main()
 {
 	printf("Hello World!\n");
 
-	CMatrice <int> MatriceEntier1(3, 2);
-	CMatrice <int> MatriceEntier2(MatriceEntier1);
+	CMatrice <double> MatriceEntier1(3, 2);
+	CMatrice <double> MatriceEntier2(MatriceEntier1);
 
 	for (unsigned int i = 0; i < MatriceEntier1.MTPLire_NbLigne(); i++)
 	{
@@ -22,6 +22,7 @@ int main()
 		}
 	}
 
+	
 	MatriceEntier1.MTPAfficherMatrice();
 	MatriceEntier2.MTPAfficherMatrice();
 
@@ -33,5 +34,3 @@ int main()
 	getchar();
 	return 0;
 }
-
-
